requirements.txt
Django==2.1.5
gunicorn==19.9.0
whitenoise==4.1.2
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
click==5.1
compass-stylesheets==0.12.6
django-cors-headers==3.0.2
django-mustache==1.1.1
djangorestframework==3.9.4
djangorestframework-api-key==1.4.0
FormEncode==1.3.1
html-json-forms==1.1.0
idna==2.8
linecache2==1.0.0
Markdown==3.1.1
natural-keys==1.5.0.post2
Pillow==6.1.0
psycopg2-binary==2.8.3
py-mini-racer==0.1.18
PyBabeljs==0.0.8
pycodestyle==2.5.0
pyScss==1.3.5
pystache==0.5.4
pytz==2019.1
pyxform==0.14.0
PyYAML==5.1.1
requests==2.22.0
requirejs==0.2.1
six==1.12.0
sqlparse==0.3.0
traceback2==1.4.0
unicodecsv==0.14.1
unittest2==1.1.0
urllib3==1.25.3
wq==1.1.0
wq.app==1.1.2
wq.core==1.1.0
wq.db==1.1.3
wq.io==1.1.0
wq.start==1.1.1
xlrd==1.2.0
xlsconv==1.1.0
